+++ packages/bruno-app/src/components/ResponseExample/ResponseExampleRequestPane/ResponseExampleMultipartFormParams/index.js
@@ -1,176 +1,18 @@
-import React, { useMemo, useCallback, useLayoutEffect, useRef, useState } from 'react';
+import React, { useMemo, useCallback } from 'react';
 import get from 'lodash/get';
 import { useDispatch, useSelector } from 'react-redux';
 import { useTheme } from 'providers/Theme';
-import { IconUpload, IconX, IconFile, IconChevronDown } from '@tabler/icons';
+import { IconUpload } from '@tabler/icons';
 import { updateResponseExampleMultipartFormParams } from 'providers/ReduxStore/slices/collections';
 import { browseFiles } from 'providers/ReduxStore/slices/collections/actions';
 import { updateTableColumnWidths } from 'providers/ReduxStore/slices/tabs';
 import mime from 'mime-types';
-import path from 'utils/common/path';
+import path, { getRelativePath } from 'utils/common/path';
 import EditableTable from 'components/EditableTable';
 import MultiLineEditor from 'components/MultiLineEditor';
 import SingleLineEditor from 'components/SingleLineEditor';
-import Dropdown from 'components/Dropdown';
-import StyledWrapper, { OverflowList } from './StyledWrapper';
-import { isWindowsOS } from 'utils/common/platform';
-
-const MIN_CHIP_W = 75;
-const CHIP_GAP = 4;
-const UPLOAD_RESERVE = 28;
-const MORE_CHIP_RESERVE = 56;
-
-const basename = (filePath) => {
-  if (!filePath) return '';
-  const separator = isWindowsOS() ? '\\' : '/';
-  return String(filePath).split(separator).pop() || String(filePath);
-};
-
-const FileChipsCell = ({ files, onRemove, onAdd, editMode }) => {
-  const containerRef = useRef(null);
-  const [visibleCount, setVisibleCount] = useState(files.length);
-
-  useLayoutEffect(() => {
-    const container = containerRef.current;
-    if (!container) return;
-    const td = container.closest('td') || container.parentElement;
-    if (!td) return;
-
-    const compute = () => {
-      const tdStyle = window.getComputedStyle(td);
-      const padX = parseFloat(tdStyle.paddingLeft) + parseFloat(tdStyle.paddingRight);
-      const total = td.clientWidth - padX;
-      if (files.length === 0) {
-        setVisibleCount(0);
-        return;
-      }
-
-      const allAtMin = files.length * MIN_CHIP_W + Math.max(0, files.length - 1) * CHIP_GAP;
-      if (allAtMin + UPLOAD_RESERVE <= total) {
-        setVisibleCount(files.length);
-        return;
-      }
-
-      const available = total - UPLOAD_RESERVE - MORE_CHIP_RESERVE;
-      const n = Math.max(0, Math.floor((available + CHIP_GAP) / (MIN_CHIP_W + CHIP_GAP)));
-      setVisibleCount(n);
-    };
-
-    compute();
-    const ro = new ResizeObserver(compute);
-    ro.observe(td);
-    return () => ro.disconnect();
-  }, [files]);
-
-  const visible = files.slice(0, visibleCount);
-  const overflow = files.slice(visibleCount);
-  const collapsed = visibleCount === 0 && files.length > 0;
-
-  const renderChip = (filePath, idx, opts = {}) => (
-    <div
-      key={`${filePath}-${idx}`}
-      className={`file-chip${opts.fullWidth ? ' file-chip-row' : ''}`}
-      title={filePath}
-    >
-      <IconFile size={14} stroke={1.5} className="file-chip-icon" />
-      <span className="file-chip-name">{basename(filePath)}</span>
-      {editMode && (
-        <button
-          type="button"
-          className="file-chip-remove"
-          onClick={(e) => {
-            e.stopPropagation();
-            onRemove(filePath);
-          }}
-          title="Remove file"
-        >
-          <IconX size={13} stroke={1.5} />
-        </button>
-      )}
-    </div>
-  );
-
-  const renderOverflowList = (list) => (
-    <OverflowList>
-      {list.map((p, i) => (
-        <div key={`o-${p}-${i}`} className="overflow-row" title={p}>
-          <IconFile size={14} stroke={1.5} className="overflow-row-icon" />
-          <span className="overflow-row-name">{basename(p)}</span>
-          {editMode && (
-            <button
-              type="button"
-              className="overflow-row-remove"
-              onClick={(e) => {
-                e.stopPropagation();
-                onRemove(p);
-              }}
-              title="Remove file"
-            >
-              <IconX size={13} stroke={1.5} />
-            </button>
-          )}
-        </div>
-      ))}
-    </OverflowList>
-  );
-
-  return (
-    <div className="file-value-cell" ref={containerRef}>
-      {collapsed ? (
-        <>
-          <Dropdown
-            placement="bottom-start"
-            appendTo={() => document.body}
-            icon={(
-              <button
-                type="button"
-                className="file-summary-chip"
-                onClick={(e) => e.stopPropagation()}
-                title={`${files.length} file${files.length > 1 ? 's' : ''}`}
-              >
-                <IconFile size={14} stroke={1.5} className="file-chip-icon" />
-                <span>{files.length} file{files.length > 1 ? 's' : ''}</span>
-                <IconChevronDown size={14} stroke={1.5} />
-              </button>
-            )}
-          >
-            {renderOverflowList(files)}
-          </Dropdown>
-          <div className="file-chips-row" />
-        </>
-      ) : (
-        <>
-          <div className="file-chips-row">
-            {visible.map((p, i) => renderChip(p, i))}
-          </div>
-          {overflow.length > 0 && (
-            <Dropdown
-              placement="bottom-end"
-              appendTo={() => document.body}
-              icon={(
-                <button
-                  type="button"
-                  className="file-more-chip"
-                  onClick={(e) => e.stopPropagation()}
-                  title={`${overflow.length} more file${overflow.length > 1 ? 's' : ''}`}
-                >
-                  +{overflow.length} more
-                </button>
-              )}
-            >
-              {renderOverflowList(overflow)}
-            </Dropdown>
-          )}
-        </>
-      )}
-      {editMode && (
-        <button type="button" className="upload-btn ml-1" onClick={onAdd} title="Add files">
-          <IconUpload size={16} />
-        </button>
-      )}
-    </div>
-  );
-};
+import MultipartFileChipsCell from 'components/MultipartFileChipsCell';
+import StyledWrapper from './StyledWrapper';
 
 const ResponseExampleMultipartFormParams = ({ item, collection, exampleUid, editMode = false }) => {
   const dispatch = useDispatch();
@@ -213,7 +55,7 @@ const ResponseExampleMultipartFormParams = ({ item, collection, exampleUid, edit
         const processedPaths = filePaths.map((filePath) => {
           const collectionDir = collection.pathname;
           if (filePath.startsWith(collectionDir)) {
-            return path.relative(collectionDir, filePath);
+            return getRelativePath(collectionDir, filePath);
           }
           return filePath;
         });
@@ -223,9 +65,13 @@ const ResponseExampleMultipartFormParams = ({ item, collection, exampleUid, edit
         const existingValue = existingParam && existingParam.type === 'file' && Array.isArray(existingParam.value)
           ? existingParam.value
           : [];
+        const seen = new Set(existingValue);
         const merged = [...existingValue];
         for (const p of processedPaths) {
-          if (!merged.includes(p)) merged.push(p);
+          if (!seen.has(p)) {
+            seen.add(p);
+            merged.push(p);
+          }
         }
 
         let updatedParams;
@@ -344,7 +190,7 @@ const ResponseExampleMultipartFormParams = ({ item, collection, exampleUid, edit
         const fileList = row.type === 'file' ? getFileList(value) : [];
         if (fileList.length > 0) {
           return (
-            <FileChipsCell
+            <MultipartFileChipsCell
               files={fileList}
               onRemove={(filePath) => handleRemoveFile(row, filePath)}
               onAdd={() => handleBrowseFiles(row, onChange)}
